Desktop app/requirements.txt
flet==0.25.2
flet-desktop==0.25.2
python-dotenv>=1.0.0
pyinstaller==6.11.1
requests>=2.28.0
psutil>=5.9.0
asyncio>=3.4.3

